feat: persist data with localStorage hooks
